refactor(dock): reduce facade header
